refactor: separate DesktopPortal classes
- separate UBDesktopPortal into UBScreenshotDesktopPortalWrapper
  and UBScreenCastDesktopPortalWrapper
- extract token generator to UBDesktopPortalTokenGenerator
- move showGlassPane to UBApplicationController
- move Desktop Portal classes to frameworks/linux subdirectory
