+++ src/frameworks/linux/UBScreenCastDesktopPortalWrapper.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2015-2025 Département de l'Instruction Publique (DIP-SEM)
+ * Copyright (C) 2015-2026 Département de l'Instruction Publique (DIP-SEM)
  *
  * This file is part of OpenBoard.
  *
@@ -20,7 +20,7 @@
  */
 
 
-#include "UBDesktopPortal.h"
+#include "UBScreenCastDesktopPortalWrapper.h"
 
 #include <QDBusArgument>
 #include <QDBusInterface>
@@ -29,81 +29,22 @@
 #include <QDBusConnection>
 #include <QDBusMetaType>
 #include <QDBusVariant>
-#include <QGuiApplication>
-#include <QClipboard>
-#include <QTimer>
 #include <QMetaType>
 #include <QDebug>
 
-#include "board/UBBoardView.h"
-#include "core/UBApplication.h"
-#include "core/UBApplicationController.h"
 #include "core/UBSettings.h"
-#include "desktop/UBDesktopAnnotationController.h"
-#include "frameworks/UBPlatformUtils.h"
+#include "frameworks/linux/UBDesktopPortalTokenGenerator.h"
 
 
 enum : uint { MONITOR = 1, WINDOW = 2, VIRTUAL = 4 } SourceType;
 enum : uint { HIDDEN = 1, EMBEDDED = 2, METADATA = 4 } CursorMode;
 enum : uint { TRANSIENT = 0, APPLICATION = 1, PERSISTENT = 2 } PersistMode;
 
-Q_DECLARE_METATYPE(UBDesktopPortal::Stream)
-Q_DECLARE_METATYPE(UBDesktopPortal::Streams)
-
-QPixmap UBDesktopPortal::loadScreenshotFromUri(const QUrl& uri) const
-{
-    if (uri.isEmpty())
-    {
-        return {};
-    }
-
-    QFile file(uri.toLocalFile());
-
-    if (!file.exists())
-    {
-        qDebug() << "Screenshot image file does not exist" << uri;
-        return {};
-    }
-
-    QPixmap pixmap{file.fileName()};
-    file.remove();
-    qDebug() << "DesktopPortal: loaded screenshot file" << uri;
-    return pixmap;
-}
-
-QPixmap UBDesktopPortal::readClipboardScreenshot() const
-{
-    auto clipboard = QGuiApplication::clipboard();
-
-    if (!clipboard)
-    {
-        qDebug() << "DesktopPortal: no clipboard available";
-        return {};
-    }
-
-    QImage cbImage = clipboard->image();
-
-    if (!cbImage.isNull())
-    {
-        QPixmap fromImage = QPixmap::fromImage(cbImage);
-        qDebug() << "DesktopPortal: clipboard image found" << "size" << fromImage.size();
-        return fromImage;
-    }
-
-    QPixmap cbPixmap = clipboard->pixmap();
-
-    if (!cbPixmap.isNull())
-    {
-        qDebug() << "DesktopPortal: clipboard pixmap found" << "size" << cbPixmap.size();
-        return cbPixmap;
-    }
-
-    qDebug() << "DesktopPortal: clipboard empty";
-    return {};
-}
+Q_DECLARE_METATYPE(UBScreenCastDesktopPortalWrapper::Stream)
+Q_DECLARE_METATYPE(UBScreenCastDesktopPortalWrapper::Streams)
 
 
-QDBusArgument &operator << (QDBusArgument &arg, const UBDesktopPortal::Stream &stream)
+QDBusArgument &operator << (QDBusArgument &arg, const UBScreenCastDesktopPortalWrapper::Stream &stream)
 {
     arg.beginStructure();
     arg << stream.node_id;
@@ -121,7 +62,7 @@ QDBusArgument &operator << (QDBusArgument &arg, const UBDesktopPortal::Stream &s
     return arg;
 }
 
-const QDBusArgument &operator >> (const QDBusArgument &arg, UBDesktopPortal::Stream &stream)
+const QDBusArgument &operator >> (const QDBusArgument &arg, UBScreenCastDesktopPortalWrapper::Stream &stream)
 {
     arg.beginStructure();
     arg >> stream.node_id;
@@ -144,9 +85,9 @@ const QDBusArgument &operator >> (const QDBusArgument &arg, UBDesktopPortal::Str
     return arg;
 }
 
-QDBusArgument &operator << (QDBusArgument &arg, const UBDesktopPortal::Streams &streams)
+QDBusArgument &operator << (QDBusArgument &arg, const UBScreenCastDesktopPortalWrapper::Streams &streams)
 {
-    arg.beginArray(qMetaTypeId<UBDesktopPortal::Stream>());
+    arg.beginArray(qMetaTypeId<UBScreenCastDesktopPortalWrapper::Stream>());
     for (const auto &stream : streams)
     {
         arg << stream;
@@ -155,14 +96,13 @@ QDBusArgument &operator << (QDBusArgument &arg, const UBDesktopPortal::Streams &
     return arg;
 }
 
-
-const QDBusArgument &operator >> (const QDBusArgument &arg, UBDesktopPortal::Streams &streams)
+const QDBusArgument &operator >> (const QDBusArgument &arg, UBScreenCastDesktopPortalWrapper::Streams &streams)
 {
     streams.clear();
     arg.beginArray();
     while (!arg.atEnd())
     {
-        UBDesktopPortal::Stream stream;
+        UBScreenCastDesktopPortalWrapper::Stream stream;
         arg >> stream;
         streams.append(stream);
     }
@@ -170,14 +110,14 @@ const QDBusArgument &operator >> (const QDBusArgument &arg, UBDesktopPortal::Str
     return arg;
 }
 
-UBDesktopPortal::UBDesktopPortal(QObject* parent)
+UBScreenCastDesktopPortalWrapper::UBScreenCastDesktopPortalWrapper(QObject* parent)
     : QObject{parent}
 {
     qDBusRegisterMetaType<Stream>();
     qDBusRegisterMetaType<Streams>();
 }
 
-UBDesktopPortal::~UBDesktopPortal()
+UBScreenCastDesktopPortalWrapper::~UBScreenCastDesktopPortalWrapper()
 {
     if (mScreencastPortal)
     {
@@ -185,57 +125,7 @@ UBDesktopPortal::~UBDesktopPortal()
     }
 }
 
-void UBDesktopPortal::grabScreen(QScreen* screen, bool interactive)
-{
-    mScreenRect = screen->geometry();
-    mInteractiveScreenshot = interactive;
-
-    qDebug() << "DesktopPortal: grabScreen"
-             << "screen" << (screen ? screen->name() : QStringLiteral("<null>"))
-             << "geom" << mScreenRect
-             << "interactive" << interactive;
-
-    QDBusInterface screenshotPortal("org.freedesktop.portal.Desktop", "/org/freedesktop/portal/desktop", "org.freedesktop.portal.Screenshot");
-
-    if (screenshotPortal.isValid())
-    {
-        QMap<QString, QVariant> options;
-        options["handle_token"] = createRequestToken();
-        options["interactive"] = mInteractiveScreenshot; // required by some portals (e.g. GNOME)
-        QDBusReply<QDBusObjectPath> reply = screenshotPortal.call("Screenshot", "", options);
-        QDBusObjectPath objectPath = reply.value();
-        QString path = objectPath.path();
-        qDebug() << "DesktopPortal: Screenshot reply valid" << reply.isValid() << "path" << path << "options" << options;
-
-        if (!reply.isValid())
-        {
-            qWarning() << "DesktopPortal: Screenshot call failed" << reply.error();
-            emit screenGrabbed(QPixmap{});
-            return;
-        }
-
-        const bool connected = QDBusConnection::sessionBus().connect(
-                            "",
-                            path,
-                            "org.freedesktop.portal.Request",
-                            "Response",
-                            "ua{sv}",
-                            this,
-                            SLOT(handleScreenshotResponse(uint,QMap<QString,QVariant>)));
-
-        if (!connected)
-        {
-            qWarning() << "DesktopPortal: failed to connect to screenshot response signal";
-        }
-    }
-    else
-    {
-        qDebug() << "No valid screenshot portal";
-        emit screenGrabbed(QPixmap{});
-    }
-}
-
-void UBDesktopPortal::startScreenCast(bool withCursor)
+void UBScreenCastDesktopPortalWrapper::startScreenCast(bool withCursor)
 {
     mWithCursor = withCursor;
 
@@ -247,9 +137,9 @@ void UBDesktopPortal::startScreenCast(bool withCursor)
     }
 
     // Create ScreenCast session
-    QString requestToken = createRequestToken();
-    QMap<QString, QVariant> options;
-    options["session_handle_token"] = createSessionToken();
+    QString requestToken = UBDesktopPortalTokenGenerator::generateToken();
+    QVariantMap options;
+    options["session_handle_token"] = UBDesktopPortalTokenGenerator::generateToken();
     options["handle_token"] = requestToken;
 
     QDBusConnection::sessionBus().connect("", mRequestPath + requestToken, "org.freedesktop.portal.Request", "Response", "ua{sv}", this,
@@ -266,7 +156,7 @@ void UBDesktopPortal::startScreenCast(bool withCursor)
     }
 }
 
-void UBDesktopPortal::stopScreenCast()
+void UBScreenCastDesktopPortalWrapper::stopScreenCast()
 {
     if (mSession.isEmpty())
     {
@@ -289,70 +179,7 @@ void UBDesktopPortal::stopScreenCast()
     mSession.clear();
 }
 
-void UBDesktopPortal::handleScreenshotResponse(uint code, const QMap<QString, QVariant>& results)
-{
-    qDebug() << "DesktopPortal: handleScreenshotResponse code" << code << "results" << results;
-
-    QPixmap pixmap;
-    auto finalizeScreenshot = [this](const QPixmap& pixmap){
-        if (pixmap.isNull())
-        {
-            qDebug() << "DesktopPortal: screenshot pixmap still null, aborting";
-            emit screenGrabbed(QPixmap{});
-            return;
-        }
-
-        QPixmap screenshot;
-
-        if (mInteractiveScreenshot)
-        {
-            // portal already applied the selection
-            screenshot = pixmap;
-        }
-        else
-        {
-            QRect targetRect = mScreenRect;
-            // NOTE have to disable this again.
-            // On KDE, the screenshot contains all screens, and I have to cut the desired screen out of this.
-            // What happens for non-interactive screenshots on GNOME? Is this important at all?
-//            targetRect.moveTo(0, 0); // portal images are local to the grab
-            targetRect &= pixmap.rect();
-            screenshot = pixmap.copy(targetRect);
-            qDebug() << "DesktopPortal: cropping screenshot" << "targetRect" << targetRect;
-        }
-
-        qDebug() << "DesktopPortal: emitting screenshot"
-                 << "interactive" << mInteractiveScreenshot
-                 << "source size" << pixmap.size()
-                 << "final size" << screenshot.size();
-        emit screenGrabbed(screenshot);
-    };
-
-    const QUrl uri(results.value("uri").toUrl());
-    qDebug() << "DesktopPortal: screenshot URI" << uri;
-
-    pixmap = loadScreenshotFromUri(uri);
-
-    // GNOME may only place the shot on the clipboard when capturing a full screen
-    if (pixmap.isNull())
-    {
-        pixmap = readClipboardScreenshot();
-    }
-
-    if (pixmap.isNull())
-    {
-        // Clipboard might not be populated yet on some portals (e.g. GNOME fullscreen); retry shortly.
-        QTimer::singleShot(500, this, [this, finalizeScreenshot](){
-            QPixmap delayed = readClipboardScreenshot();
-            finalizeScreenshot(delayed);
-        });
-        return;
-    }
-
-    finalizeScreenshot(pixmap);
-}
-
-void UBDesktopPortal::handleCreateSessionResponse(uint response, const QVariantMap& results)
+void UBScreenCastDesktopPortalWrapper::handleCreateSessionResponse(uint response, const QVariantMap& results)
 {
     if (response != 0)
     {
@@ -372,8 +199,8 @@ void UBDesktopPortal::handleCreateSessionResponse(uint response, const QVariantM
     }
 
     // Select sources
-    QString requestToken = createRequestToken();
-    QMap<QString, QVariant> options;
+    QString requestToken = UBDesktopPortalTokenGenerator::generateToken();
+    QVariantMap options;
     options["multiple"] = false;
     options["types"] = MONITOR;
     options["cursor_mode"] = mWithCursor ? EMBEDDED : HIDDEN;
@@ -406,7 +233,7 @@ void UBDesktopPortal::handleCreateSessionResponse(uint response, const QVariantM
     }
 }
 
-void UBDesktopPortal::handleSelectSourcesResponse(uint response, const QVariantMap& results)
+void UBScreenCastDesktopPortalWrapper::handleSelectSourcesResponse(uint response, const QVariantMap& results)
 {
     Q_UNUSED(results);
 
@@ -426,7 +253,7 @@ void UBDesktopPortal::handleSelectSourcesResponse(uint response, const QVariantM
     }
 
     // Start ScreenCast
-    QString requestToken = createRequestToken();
+    QString requestToken = UBDesktopPortalTokenGenerator::generateToken();
     QMap<QString, QVariant> options;
     options["handle_token"] = requestToken;
 
@@ -447,7 +274,7 @@ void UBDesktopPortal::handleSelectSourcesResponse(uint response, const QVariantM
     showGlassPane(false);
 }
 
-void UBDesktopPortal::handleStartResponse(uint response, const QVariantMap& results)
+void UBScreenCastDesktopPortalWrapper::handleStartResponse(uint response, const QVariantMap& results)
 {
     // Show annotation drawing view in desktop mode after portal dialog was closed
     showGlassPane(true);
@@ -500,7 +327,7 @@ void UBDesktopPortal::handleStartResponse(uint response, const QVariantMap& resu
     }
 
     // Open PipeWire Remote
-    QMap<QString, QVariant> options;
+    QVariantMap options;
     const QDBusReply<QDBusUnixFileDescriptor> reply = portal->call("OpenPipeWireRemote", QDBusObjectPath(mSession), options);
 
     if (!reply.isValid())
@@ -515,7 +342,7 @@ void UBDesktopPortal::handleStartResponse(uint response, const QVariantMap& resu
     emit streamStarted(fd, stream.node_id);
 }
 
-QDBusInterface* UBDesktopPortal::screencastPortal()
+QDBusInterface* UBScreenCastDesktopPortalWrapper::screencastPortal()
 {
     if (!mScreencastPortal)
     {
@@ -543,33 +370,4 @@ QDBusInterface* UBDesktopPortal::screencastPortal()
 
     emit screenCastAborted();
     return nullptr;
-}
-
-QString UBDesktopPortal::createSessionToken() const
-{
-    static int sessionTokenCounter = 0;
-
-    sessionTokenCounter += 1;
-    return QString("obsess%1").arg(sessionTokenCounter);
-}
-
-QString UBDesktopPortal::createRequestToken() const
-{
-    static int requestTokenCounter = 0;
-
-    requestTokenCounter += 1;
-    return QString("obreq%1").arg(requestTokenCounter);
-}
-
-void UBDesktopPortal::showGlassPane(bool show) const
-{
-    if (UBApplication::applicationController->isShowingDesktop())
-    {
-        UBApplication::applicationController->uninotesController()->drawingView()->setVisible(show);
-
-        if (show)
-        {
-            UBPlatformUtils::keepOnTop();
-        }
-    }
 }
